too many chnages to mention. ive been slacking on pushing and just doing everything locally

+++ src/user_manager.py
@@ -2,10 +2,10 @@
 import os
 import pickle
 from pathlib import Path
-from typing import List
+from typing import Any, List, Dict
 from copy import deepcopy as copy
-
 from src.facts_maker import FactsMaker
+
 
 
 class UserManager:
@@ -13,57 +13,77 @@ class UserManager:
 
     def __init__(self):
 
-        #self._data: dict = {}
-
-        self._inc_nums = {}
-        self._inc_oprs = {}
-        
+        self.user_name: str
+        self.inc_nums: dict
+        self.inc_oprs: dict
+        self.inc_ptrns: dict
+        self.num_facts: int
+        self.inc_timers: bool
+        self.inc_untimed: bool
+        self.timer_vals: List[int]
+        self.facts: Dict[int, Dict[str, Any]]
+        self.user_loaded: bool = False
 
         self._save_dir: Path
         self._user_path: Path
-        self._user_loaded = False
 
 
-    def create_user(self, user: str):
+    def create_new_user(self, user: str):
 
-        settings = {
-            'inc_nums': {k: True for k in range(13)},
-            'inc_oprs': {'add': True, 'sub': True, 'mul': True, 'div': True},
-            'inc_ptrns': {'reversed': True, 'mixed_unknowns': True},
-            'num_facts': 25,
-            'inc_timers': True,
-            'inc_untimed': True,
-            'timers': [8, 4] }
-        self._data = {
-            'user_name': user,
-            'settings': settings,
-            'facts': FactsMaker().get_facts() }
-        
-        self._save_user_data()
-        self.load_saved_user(user)
+        self.user_name = user
+        self.inc_nums = {k: True for k in range(13)}
+        self.inc_oprs = {'add': True, 'sub': True, 'mul': True, 'div': True}
+        self.inc_ptrns = {'reversed': True, 'mixed_unknowns': True}
+        self.num_facts = 25
+        self.inc_timers = False
+        self.inc_untimed = True
+        self.timer_vals = [9, 6, 3]
+        self.facts = FactsMaker().get_facts()
+        self.user_loaded = True
 
-
-    def _save_user_data(self):
-
-        try:
-            self._set_path_attributes(self._data['user_name'])
-        except KeyError:
-            raise ValueError("call create_user() or load_saved_user() first")
-
-        os.makedirs(self._save_dir, exist_ok=True)
-        with open(self._user_path, 'wb') as f:
-            pickle.dump(self._data, f)
+        self.save_user_data()
 
 
     def load_saved_user(self, user: str):
 
         self._set_path_attributes(user)
         if not self._user_path.exists():
-            raise FileNotFoundError(f"User does not exist at {self._user_path}")
+            raise FileNotFoundError(f"user does not exist at {self._user_path}")
         with open(self._user_path, 'rb') as file:
-            self._data = pickle.load(file)
+            user_data = pickle.load(file)
 
-        self._user_loaded = True
+        self.user_name = user_data['user_name']
+        self.inc_nums = user_data['inc_nums']
+        self.inc_oprs = user_data['inc_oprs']
+        self.inc_ptrns = user_data['inc_ptrns']
+        self.num_facts = user_data['num_facts']
+        self.inc_timers = user_data['inc_timers']
+        self.inc_untimed = user_data['inc_untimed']
+        self.timer_vals = user_data['timer_vals']
+        self.facts = user_data['facts']
+        self.user_loaded = True
+
+
+    def save_user_data(self):
+
+        self._set_path_attributes(self.user_name)
+        if not self.user_loaded:
+            raise ValueError("no user loaded, cannot save")
+
+        user_data = {
+            'user_name': self.user_name,
+            'inc_nums': self.inc_nums,
+            'inc_oprs': self.inc_oprs,
+            'inc_ptrns': self.inc_ptrns,
+            'num_facts': self.num_facts,
+            'inc_timers': self.inc_timers,
+            'inc_untimed': self.inc_untimed,
+            'timer_vals': self.timer_vals,
+            'facts': self.facts }
+
+        os.makedirs(self._save_dir, exist_ok=True)
+        with open(self._user_path, 'wb') as f:
+            pickle.dump(user_data, f)
 
 
     def _set_path_attributes(self, user_name: str):
@@ -73,36 +93,12 @@ class UserManager:
         self._user_path = Path.joinpath(self._save_dir, f'{user_name}.pkl')
 
 
-    def save_new_username(self, new_username: str):
+    def delete_current_user(self):
+
+        if not self.user_loaded:
+            raise ValueError("no user loaded, cannot delete")
 
         os.remove(self._user_path)
-        self._data['user_name'] = new_username
-        self._save_user_data()
-
-
-    def save_new_settings(self, settings: dict):
-
-        self._data['settings'] = copy(settings)
-        self._save_user_data()
-
-
-    def save_new_facts(self, facts: List[dict]):
-
-        self._data['facts'] = copy(facts)
-        self._save_user_data()
-
-
-    def delete_user(self):
-
-        os.remove(self._user_path)
-        self._user_loaded = False
-        self._data = {}
-
-
-    @property
-    def data(self):
-
-        return copy(self._data)
 
 
     @staticmethod
@@ -115,80 +111,3 @@ class UserManager:
             return []
 
         return [f.stem for f in save_dir.glob('*.pkl') if f.is_file()]
-    
-
-    @property
-    def user_loaded(self) -> bool:
-
-        return self._user_loaded
-
-
-    @property
-    def include_timers(self) -> bool:
-
-        return self._data['settings']['inc_timers']
-    
-
-    # @property
-    # def settings(self) -> dict:
-
-    #     if self._ud is None:
-    #         raise Exception("call load_data() first")
-
-    #     return self._ud['settings']
-
-
-    # @property
-    # def math_facts(self) -> List[dict]:
-        
-    #     if self._ud is None:
-    #         raise Exception("call load_data() first")
-        
-    #     return self._ud['facts']
-        
-
-    # @property
-    # def num_facts(self) -> int:
-
-    #     # fill this in later
-    #     return 20
-
-
-    # @property
-    # def times(self) -> List[int]:
-
-    #     # fill this in later
-    #     return []
-
-
-    # @property
-    # def exclude(self) -> List[int]:
-
-    #     # fill this in later
-    #     return [0]
-
-    
-    
-
-    # def _on_create_clicked(self):
-
-    #     cbbox_val = self._combobox.get()
-
-    #     errs = []
-    #     if not cbbox_val[0].isalpha():
-    #         errs.append("start with a letter")            
-    #     if len(cbbox_val) < 3:
-    #         errs.append("be at least three characters long")
-    #     whitelist = string.ascii_letters + string.digits + '_'
-    #     if not all(c in whitelist for c in cbbox_val):
-    #         errs.append("contain only letters, digits, and underscores")
-    #     if errs:
-    #         message = "Username should:\n" + "\n".join(f"  - {err}" for err in errs)
-    #         self.option_add("*Dialog.msg.font", self._small_font)
-    #         self.option_add("*Dialog.msg.wrapLength", "0")
-    #         messagebox.showinfo("Requirements", message)
-    #         return
-        
-    #     self._user.create_user(cbbox_val)
-    #     self._combobox['values'] = self._user.get_usernames()
-    #     self._on_user_update()
